% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simcor.R
\name{noisecor}
\alias{noisecor}
\title{Add noise to any user specified correlation matrix}
\usage{
noisecor(cormat, epsilon = 0.01, eidim = 2)
}
\arguments{
\item{cormat}{the correlation matrix to which noise is to be added.}

\item{epsilon}{epsilon maximum entry-wise random noise.}

\item{eidim}{dimension of the noise.}
}
\value{
The noise corrupted correlation matrix.
}
\description{
From Appendix A2. This is the core algorithm for adding noise to an
arbitrary positive semi-definite matrix. The choice of eidim (M in the paper)
provides flexibilty. If ei (M) is small (between 2 and 5) then many of the dot
product terms will be large, yielding a very noisy coefficient matrix. If M is large
the computation is still cheap. Other choices are discussed in the paper. This version
has been modified to remove loops.
}
\examples{
 noise.iden <- noisecor(diag(15), epsilon = .5, eidim = 2)
}
