% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simcor.R
\name{simcor.H}
\alias{simcor.H}
\title{Simulating the Hub Matrix (entries filled in using Toeplitz structure)}
\usage{
simcor.H(k = 6, size = c(10, 5, 8, 7, 15, 50), rho = rbind(c(0.9, 0.7),
  c(0.7, 0.7), c(0.7, 0.2), c(0.5, 0.3), c(0.9, 0.85), c(0.3, 0.2)),
  power = 1, epsilon = 0.08, eidim = 2)
}
\arguments{
\item{k}{is the number of groups}

\item{size}{is a vector of length k specifying the size of each group}

\item{rho}{is a vector of length k specifying base correlation values}

\item{power}{= 2 makes the correlations stay high, = 0.5 makes the correlations descent rapidly}

\item{epsilon}{<- (1-min(rho) - 0.75*min(tau) ) - .01 - maximum entry-wise random noise.}

\item{eidim}{is the space from which the noise is generated, the smaller the more noise}
}
\description{
Simulating the Hub Matrix (entries filled in using Toeplitz structure)
}
\examples{
# Figure 2 in the paper
vw <- function(Im)
{
image(Im[, ncol(Im):1], zlim=c(-0.4, 1), col=rev(heat.colors(64)))
}
k<-3
sz <- c(100, 50, 80)
rho <- matrix(c(0.7, 0.7, 0.4, 0, 0, 0), nrow=3)
epsilon <- 0.23
hTC1 <- simcor.H(k=k, size=sz, rho=rho, power=1, epsilon=0.23, eidim=2)
vw(hTC1)
rho <- matrix(c(0.7, 0.7, 0.4, 0.5, 0.6, 0.2), nrow=3)

hTC2 <- simcor.H(k=k, size=sz, rho=rho, power=1, epsilon=0.29, eidim=2)
vw(hTC2)

hTC3 <- simcor.H(k=k, size=sz, rho=rho, power=1, epsilon=0.29, eidim=25)
vw(hTC3)

hTC4 <- simcor.H(k=k, size=sz, rho=rho, power=1, epsilon=0.1, eidim=2)
vw(hTC4)
hTC5 <- simcor.H(k=k, size=sz, rho=rho, power=1, epsilon=0.25, eidim=2)
vw(hTC5)
rho <- matrix(c(0.8, 0.75, 0.7, 0, 0, 0), nrow=3)
hTC6 <- simcor.H(k=k, size=sz, rho=rho, power=1, epsilon=0.19, eidim=2)
vw(hTC6)
}
