% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simcor.R
\name{simcor}
\alias{simcor}
\title{Block correlation noise matrix}
\usage{
simcor(k = 6, size = c(10, 5, 8, 2, 15, 50), rho = c(0.7, 0.7, 0.5, 0.9,
  0.85, 0.4), delta = 0.39, epsilon = 0.99 - max(rho), eidim = 2)
}
\arguments{
\item{k}{Number of groups.}

\item{size}{Vector length k specifying the size of each group}

\item{epsilon}{Why is this an argument}

\item{eidim}{passed to noise corr (or should be)}
}
\description{
Block correlation noise matrix
}
