% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simcor.R
\name{simcorTop}
\alias{simcorTop}
\title{Toeplitz correlation structure}
\usage{
simcorTop(k = 6, size = c(10, 5, 8, 2, 15, 50), rho = c(0.7, 0.7, 0.5,
  0.9, 0.85, 0.4), epsilon = 0.01, eidim = 2)
}
\arguments{
\item{k}{is the number of groups}

\item{size}{is a vector of length k specifying the size of each group}

\item{rho}{is a vector of length k specifying base correlation values}

\item{eidim}{is the space from which the noise is generated, the smaller the more noise}

\item{maximum}{entry-wise random noise.}
}
\description{
Toeplitz correlation structure
}
\details{
This function calculates an AR(1) Toeplitz matrix
with a block diagonal structure.  The size and base correlation for each
block is user specified
}
